requirements.txt
Flask==3.1.2
Flask-CORS==6.0.2
requests==2.32.5
validators==0.35.0
sendgrid==6.11.0
python-dotenv==1.0.0
